requirements.txt
paho-mqtt>=2.1.0
dash>=2.14.0
pandas>=2.0.0
plotly>=5.18.0